requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-google-genai
PyPDF2
faiss-cpu
python-dotenv
huggingface_hub
